+++ mpfs-rpmsg-freertos/src/middleware/rpmsg/rpmsg_lite/porting/environment/rpmsg_env_freertos.c
@@ -297,7 +297,7 @@ void env_wmb(void)
  *
  * @param address
  */
-uint32_t env_map_vatopa(void *address)
+uint64_t env_map_vatopa(void *address)
 {
     return platform_vatopa(address);
 }
@@ -307,7 +307,7 @@ uint32_t env_map_vatopa(void *address)
  *
  * @param address
  */
-void *env_map_patova(uint32_t address)
+void *env_map_patova(uint64_t address)
 {
     return platform_patova(address);
 }
